Issue-#59: Simplify the entry point name -> MoreFluents -> Fluents, since "MoreFluents" should be used the best.
