+++ .idea/jarRepositories.xml
@@ -7,6 +7,11 @@
       <option name="url" value="https://repo.maven.apache.org/maven2" />
     </remote-repository>
     <remote-repository>
+      <option name="id" value="rakuten" />
+      <option name="name" value="Rakuten Internal" />
+      <option name="url" value="http://gsp-master-nexus1.gsp.prod.rsc.local:8081/repository/public" />
+    </remote-repository>
+    <remote-repository>
       <option name="id" value="central" />
       <option name="name" value="Maven Central repository" />
       <option name="url" value="https://repo1.maven.org/maven2" />
@@ -15,6 +20,11 @@
       <option name="id" value="jboss.community" />
       <option name="name" value="JBoss Community repository" />
       <option name="url" value="https://repository.jboss.org/nexus/content/repositories/public/" />
+    </remote-repository>
+    <remote-repository>
+      <option name="id" value="central" />
+      <option name="name" value="Central Repository" />
+      <option name="url" value="http://gsp-master-nexus1.gsp.prod.rsc.local:8081/repository/public" />
     </remote-repository>
   </component>
 </project>
