+++ src/test/java/com/github/dakusui/pcond/ut/fluent/MoreFluentStreamTest.java
@@ -4,8 +4,8 @@ import org.junit.Test;
 
 import java.util.stream.Stream;
 
-import static com.github.dakusui.pcond.fluent.MoreFluents.assertWhen;
-import static com.github.dakusui.pcond.fluent.MoreFluents.valueOf;
+import static com.github.dakusui.pcond.fluent.Fluents.assertWhen;
+import static com.github.dakusui.pcond.fluent.Fluents.valueOf;
 import static com.github.dakusui.pcond.forms.Predicates.isEqualTo;
 import static com.github.dakusui.pcond.forms.Predicates.isNotNull;
